refactor(505): rework after review
