+++ src/common/components/ImageUpload/ImageUpload.vue
@@ -31,6 +31,15 @@ const imageUpload = useImageUpload()
 
 const ACCEPTED_FILE_TYPES = ['image/jpg', 'image/jpeg', 'image/png', 'application/pdf']
 
+const errorId = 'image-upload-error'
+const hintId = 'image-upload-hint'
+
+const describedBy = computed(() => {
+  return imageUpload.error.value
+    ? `${hintId} ${errorId}`
+    : hintId
+})
+
 /**
  * Method exectuted on file update
  * We only keep the first file if a list is provided
@@ -46,6 +55,7 @@ async function onUpdateImage (files: FileList) {
 
 <template>
   <AvFileUpload
+    :aria-describedby="describedBy"
     :error="imageUpload.error.value"
     :valid-message="imageUpload.valid.value"
     :accept="ACCEPTED_FILE_TYPES"
@@ -60,18 +70,29 @@ async function onUpdateImage (files: FileList) {
     </template>
     <span class="b2-bold">{{ imageUpload.name.value }}</span>
     <template #hint>
-      <span class="caption-light">
-        {{ t('student.widgets.overview.updateProfileDrawer.pictures.format.filesIndication') }}
+      <span
+        :id="hintId"
+        class="caption-light"
+      >
+        {{ t('global.information.imageUpload.filesIndication') }}
         <span class="caption-bold">
-          {{ t('student.widgets.overview.updateProfileDrawer.pictures.format.filesTypes') }}
+          {{ t('global.information.imageUpload.filesTypes') }}
         </span>
-        {{ t('student.widgets.overview.updateProfileDrawer.pictures.format.sizeIndication') }}
+        {{ t('global.information.imageUpload.sizeIndication') }}
         <span class="caption-bold">
-          {{ t('student.widgets.overview.updateProfileDrawer.pictures.format.size') }}
+          {{ t('global.information.imageUpload.size') }}
         </span>
       </span>
     </template>
   </AvFileUpload>
+  <template v-if="imageUpload.error.value">
+    <span
+      :id="errorId"
+      class="fr-sr-only"
+    >
+      {{ imageUpload.error.value }}
+    </span>
+  </template>
 </template>
 
 <style lang="scss" scoped>
