+++ src/features/student/components/widgets/StudentOverviewWidget/components/UpdateProfileDrawer/types.ts
@@ -1,0 +1,10 @@
+export interface UpdateProfileDrawerForm {
+  email: string
+  firstname: string
+  lastname: string
+  bio: string
+  coverPicture: string
+  profilePicture: string
+}
+
+export type FormKeys = keyof UpdateProfileDrawerForm
