+++ src/features/student/locales/fr.json
@@ -322,12 +322,6 @@
           },
           "pictures": {
             "banner": "Couverture",
-            "format": {
-              "filesIndication": "Format : ",
-              "filesTypes": "PDF, JPG, JPEG, PNG •",
-              "sizeIndication": "Poids : ",
-              "size": "10 Mo max"
-            },
             "picture": "Photo de profil"
           },
           "title": "Modifier mon profil"
